+++ src/__init__.py
@@ -1,3 +1,5 @@
 from .field_types import\
 	PdfFieldType,\
 	get_field_type
+
+from .radio_btn_group import RadioBtnGroup
